Clean up the data model doc and move it to developer documentation (#89)
After some discussion, it was suggested that the user documentation be
much more focused on technical descriptions of each aspect of the LFRic
system and more compact, and that the original data model that described
the PSyKAl architecture and how LFRic worked alongside PSyclone should
be moved to developer documentation (it was agreed that it may not be an
ideal location but it was better than either user documentation or the
getting going section).

This pull request makes those changes. It also makes changes to what was
called lfric_datamodel:
* In reality, the document focuses on fields and the PSyKAl
architecture, and the detail of the core data model that was influenced
by the architecture and the requirements of GungHo, so the name has been
changed.
 *  Some of the introductory text has been changed.
* Several extraneous bits of the document have been removed as they are
best incorporated elsewhere.
* In working through the document, a few other minor changes have been
done.